pyproject.toml
[project]
name = "nginx-parser"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Eldar Mamedov", email = "bumbleboo85@gmail.com" }
]
dependencies = [
    "django-ninja>=1.3.0",
    "pydantic-settings>=2.4.0",
    "psycopg[binary,pool,c]>=3.2.1",
    "polars>=1.5.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6.2",
    "mypy>=1.11.1",
    "faker>=28.0.0",
    "pytest>=8.3.2",
    "pytest-django>=4.8.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/nginx_parser"]



[tool.ruff]
# match black
line-length = 120
target-version = "py312"
lint.select = [
    "ALL", # include all the rules, including new ones
]
lint.ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    # "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D205",
    "D212",
    "D203",
    "TRY003", # external messages in exceptions are too verbose
    "INP001",
    "ERA001",
    "RUF002",
    "RUF003",
    "EM101",
    "EM102",
    "B008",
    "G004",
    "DTZ005",
    "PLC0105",
    "A002",
    "PGH003",
    "FBT001",
    "FBT002",
    "S101",
    "RUF001",
    "PLR0913",
    "D202",
    "PLW0120",
    "RUF012",
]
exclude = ["*/migrations/*", "*/tests/*"]

[tool.mypy]
python_version = "3.12"
follow_imports = "silent"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = false
no_implicit_optional = true
strict_optional = true
files = "src"
exclude = [
    "tests/*",
    "*/migrations/*",
    "*/settings.py",
    "*/manage.py"
]


[tool.ruff.lint.per-file-ignores]
"*/models.py" = ["RUF012"]
"*/models/" = ["RUF012"]
"*/api.py" = ["ARG001"]
